requirements.txt
anthropic>=0.40.0
edge-tts>=6.1.9
moviepy>=1.0.3,<2.0.0
imageio-ffmpeg>=0.4.9
Pillow>=10.0.0
pytrends>=4.9.2
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
duckduckgo-search>=6.0.0
